lama-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.121.3
filelock==3.20.0
fire==0.5.0
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
Pillow==9.5.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
simple-lama-inpainting==0.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
termcolor==3.2.0
torch==2.9.1
torchvision==0.24.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0